MainActivity: start home if launcher's background is transparent [1/1]
PD#SWPL-4142

Problem:
it relate to google launcher, the laucher's background is transparent.
when go back to launcher and switch the window,
it will flash a black screen for a moment by android original flow,
then the WindowManager start to draw the view of Laucnher.

Solution:
start home when press BACK key in GTVS platform

Verify:
verify by einstein

Change-Id: If13acb6c3cb276ed26a499821decb813ab28b632
Signed-off-by: Lei Qian <lei.qian@amlogic.com>